+++ CHANGELOG.md
@@ -60,4 +60,24 @@ One can change this param or turn the restriction off.
 
 
 #### 0.3.1 – 04.08.2020
-Searching with gramm params fixed. 
+Searching with gramm params fixed – **Issue 3** closed.
+
+
+#### 0.3.2 – 06.08.2020
+#### Changed
+* Default filename contains letters and digits, len = 8.
+
+#### Added
+* Corpora classes to `rnc.`
+* Working with file: validating, that the Corpus type in the file
+is equal to the Corpus class type.   
+* Requesting two or more words with one str – **Issue 5** closed.
+* Deleting the example by the index – **Issue 6** closed.
+* Method `filter` to Corpus – **Issue 9** closed. 
+
+
+#### Fixed
+* Distance between words set – **Issue 7** closed. 
+* Order of texts in the ParallelCorpus – **Issue 8** closed.
+
+Other minor fixes and improvements.